refactoring, added routing
