+++ backend/Controllers/RestrictionController.js
@@ -1,0 +1,7 @@
+//Fetches all restrictions
+module.exports = {
+    getRestrictions: function(req, res) {
+        res.send("not implemented yet");
+    }
+}
+
